Split up the Catalog trait and use mockall
Replace the MockCatalog structure with mockall to avoid having to implement
dummy methods for everything we add to `Catalog.` Also split the region loading
out of the `Catalog`. This is because we need to add the region loader to the
Table structure so that it can lazily load regions at query plan time, but that
would otherwise mean a circular reference when mocking.