Use an Enum for our logical node
This means we don't have to write out implementations of everything for the
`UserDefinedLogicalNode` trait (we just have a single implementation in the
enum). It also works around a gnarly issue where, when unwrapping the node
in `plan_query`, we cast it to `dyn Any`. In that case, using `await` in the
vicinity of this captures the `dyn Any` in a Future, making that Future
unable to implement Send. This, in turn, causes an error when trying to use
the query planner in a server middleware context _in a completely different_
place, without any information as to what's causing the future to be `!Send`.

```
the trait `std::marker::Sync` is not implemented for `dyn std::any::Any`
```

see https://blog.rust-lang.org/inside-rust/2019/10/11/AsyncAwait-Not-Send-Error-Improvements.html
for more info.