Run the PG/HTTP frontends selectively in `main`
Because we prepare a vector of all futures (PG/HTTP frontends), there's a
lifetime issue with the config structs that I don't completely understand
(they basically need to live as long as the `'static` lifetime). We can't pass
them by reference to the PG/HTTP frontend because that reference might not live
as long as static, since we're doing this as a separate function that's called
from main (?)). Our `prepare_frontends` function can't own the config either
(it has to outlive it). To fix this, we pass the config to the PG/HTTP frontends
by-value and clone it.