Move postgres.rs to default.rs for better diffs
The change that follows makes a macro in `default.rs` to implement `Repository`
for an arbitrary sqlx executor. A lot of the code is the same as `postgres.rs`,
but the changes are big enough for Git to consider it as a separate file. Commit
this as a rename so that changes to `postgres.rs` have an easier time merging
into this.