Delete remains of sqlx compile-time query checking
As per https://github.com/splitgraph/seafowl/pull/34, we can't use sqlx
compile-time query checking with both SQLite and PostgreSQL at the same time. It
also can impair other devs getting started with the repo (have to maintain the
sqlx-data.json file; keeping the file up to date with sqlx requires
recompilation for some reason).

Instead, we will rely on the test coverage in the Repository trait definition,
as well as in end-to-end tests.