Add catalog-specific errors and propagate them up
It might be a better idea to merge the catalog and repository traits, as we now
have an intermediate error type (for FK/unique constraint violations) that we
don't really care about, since we immediately reinterpret them into
a catalog error depending on the context (e.g. FK violation -> table/collection
doesn't exist, unique violation -> table/collection already exists). Presumably
we'll also be converting these into `DataFusionError`s in the context (and then
to `ApiError`s in the HTTP frontend).

There's also a bit too much boilerplate `map_err` for my liking.

To keep this compiling and the tests passing, move all `unwrap()s` up to the
context level (so this will still panic if these errors are encountered, exactly
the same as currently -- but we'll be able to start removing these `unwrap()`s
one by one).