Run `reload_schema` automatically
It looks like the caller needs to run it before/after every query (that resulted
in a schema change, but maybe not the table partition change) and the actual
frontends have to run it before every query anyway. Run it automatically in
`create_logical_plan` to avoid potential confusion in tests. We'll figure out
later if it's possible to optimize it to improve latency costs.