Add support for multiple statements in `POST /q`
This is hacky because we can't plan them all at the same time, as an earlier
statement in a batch might create a table that won't be visible when we're
planning the next statement. We also need to make sure the user is only running
one read statement as a final one in the query (no support for serializing
multiple query results, just send separate read queries instead), as well as
make sure a reader still can't execute write statements.