Store full Arrow field JSON serializations
If we store just the types, we lose some vital type modifiers (nullability and,
in the case of structs/lists, the actual type of the contents of the list). To
fix this, store the full field definition (which contains the children separate
of the type).

Also add migrations to convert the "just the type" JSON objects to the "field"
JSON object by pushing down the current JSON into the `type` field (tested
manually).