Percent-decode incoming queries
headers cannot contain special characters such as newlines
(as is described in GH issue #104). The solution is to
percent-encode (aka: encodeURIComponent()) the query string.
Encoding is not mandatory, some characters like the space
are HTTP header-safe, these characters can be sent both
as they are or percent encoded (e.g. %20 for space).