Remove aliases from optimized `Update`/`Delete`s
These expressions are similar to what DataFusion uses in the `Filter` node and
not doing this seems to break partition pruning (perhaps it stops at the
`Alias` node and doesn't prone anything, didn't investigate in depth).

Copy the `ExprRewriter` visitor from
https://github.com/apache/arrow-datafusion/blob/c50573939d21de40e591c04915d41f7c46a51d0d/datafusion/expr/src/utils.rs#L384-L428
and adapt it to remove aliases from all expressions that the query optimizer
gives back to `Update`/`Delete` nodes.