Fix for servers that don't support `HEAD`
Pre-signed S3 URLs can only work with a `HEAD` or a `GET` request, but not both
at the same time. As a workaround, if we get a 403 in response to a `HEAD`,
try sending a `GET` request with a zero range. The response should have a
`Content-Range` header showing the total length of the resource, which we report
back to DataFusion.