Don't enforce a file extension when querying HTTP
If our HTTP URL has a query string, it won't end with an actual file extension.
DataFusion's `ListingTable` filters by the extension by default (useful when
scanning through a directory where some files aren't part of the table). To fix
this, disable this feature when using HTTP (since we can't scan through a
directory in HTTP anyway).

Also extend the e2e HTTP test to have a query string in it (ignored by the mock,
but exercises this workaround).