Fix `DELETE` filter plan logic with string filters
When processing a `DELETE` statement, we use `WHERE` clause to see which
partitions to scan through (vs copy directly to the new table version, since
they're definitely unaffected by the `DELETE`). Then, we construct a scan
through the remaining partitions that only keeps the rows that would remain
after the `DELETE`.

However, we invert the `WHERE` clause only in the filter expression, not the
`scan_filters` that go into the Parquet file reader. This means that the Parquet
file reader will discard partitions that definitely don't need to be deleted
(as opposed to discarding partitions that definitely don't need to be kept).