Add per-scan metrics to CachingObjectStore
The Factory now returns a callback that's supposed to build a
per-scan, per-table LogStore + a MetricsSet that it gathers. For
normal object stores, we won't collect metrics, whereas for
CachingObjectStore, we want to collect metrics and attach them
to each table scan so that we can see them in ANALYZE. To do
that, add support for cloning a CachingObjectStore with the
same cache, same file manager, same underlying client, but a
brand new MetricsSet, and expose it in our callback.