Allow running Seafowl without an object store / catalog config.
In this case, Seafowl expects the metastore and the object store settings
to be passed to it over the Flight RPC endpoint (when querying or writing)
and will error out otherwise.

This required some weird refactoring because so many things assume there is
a "default" object store and a metastore, so all SeafowlContext methods that
want to access the internal default object store for writes now do it through a method
that errors out if it's not set up. For the metastore, this change defines an
EmptyMetastore that errors if Seafowl actually tries to use it.